add to_static
